Add an option to OVERRIDE
Providing an OVERRIDE tarball allows any and everything about the image
to be customized *before* the initial commands are run. This will be
used in the gate and other CICD systems to change package sources as
well as configure options.

In infra specfically we will be disabling signature verification on the
packages and pointing the repos to use infra repos.

Regular building is 100% unaffected by this addition.

Change-Id: Ifd7f8f1724b842a396a791cf6af1b7d8848e5ff5
Depends-On: Ia664c54c2d726c03a20d8ca7f7512be3d6cfbbf9